requirements.txt
asgiref==3.2.10
Django==3.1.2
pkg-resources==0.0.0
psycopg2-binary==2.8.6
pytz==2020.1
sqlparse==0.3.1
django-crispy-forms==1.9.2